[tidb] Add tidb cdc region change unit test and polish TiKVRichParallelSourceFunction.
